new-version
1

+++ hotel/src/main/java/com/project/hotel/repository/BookingRepository.java
@@ -1,7 +1,7 @@
 package com.project.hotel.repository;
 
-import com.project.hotel.entity.Booking;
+import com.project.hotel.model.entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
+public interface BookingRepository extends JpaRepository<Booking, Long> {
 }
